feat(pluggable-widgets-commons): unify selection context value creation
