+++ packages/pluggableWidgets/datagrid-web/src/Datagrid.tsx
@@ -10,7 +10,12 @@ import {
     useFilterContext,
     useMultipleFiltering
 } from "@mendix/pluggable-widgets-commons/components/web";
-import { getGlobalSelectionContext, isAvailable, useSelectionHelper } from "@mendix/pluggable-widgets-commons";
+import {
+    getGlobalSelectionContext,
+    isAvailable,
+    useCreateSelectionContextValue,
+    useSelectionHelper
+} from "@mendix/pluggable-widgets-commons";
 import { extractFilters } from "./features/filters";
 import { useCellRenderer } from "./features/cell";
 import { getColumnAssociationProps, isSortable } from "./features/column";
@@ -121,16 +126,7 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
     const selectActionProps = useOnSelectProps(selection);
     const { selectionStatus, selectionMethod } = selectionSettings(props, selection);
 
-    const toggleSelection = selectActionProps.onSelectAll;
-    const multiSelectionStatus = selection?.type === "Multi" ? selection.selectionStatus : undefined;
-    const selectionContextValue = useMemo(() => {
-        if (multiSelectionStatus !== undefined) {
-            return {
-                status: multiSelectionStatus,
-                toggle: toggleSelection
-            };
-        }
-    }, [multiSelectionStatus, toggleSelection]);
+    const selectionContextValue = useCreateSelectionContextValue(selection);
 
     return (
         <Table
